refactor: responseHandler 분리 및 auth.service 수정
