+++ src/features/auth/actions/regist.action.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { regitsterUser } from "@/services/auth.service";
+import { registerUser } from "@/services/auth.service";
 import { RegitstRequest, ROLE } from "../types";
 import { validateRegistForm } from "../validate";
 import { BackendError } from "@/shared/types/types";
@@ -44,7 +44,7 @@ export const registAction = async (
   };
 
   try {
-    await regitsterUser(payload);
+    await registerUser(payload);
   } catch (error) {
     //백엔드 에러 코드 기반으로 매핑
     const err = error as BackendError;
